backend/requirements.txt
openai
openpyxl
requests
pandas
beautifulsoup4
numpy
uvicorn
fastapi
rich
python-dotenv
fastapi-cors
chardet
pytrends
Pillow
datasets
transformers
sentence-transformers
pinecone-client
pinecone-text
protobuf
google-generativeai
langchain_community
langchain_openai
langchain-google-genai
